spotify-playlist-importer/requirements.txt
spotipy==2.23.0
mutagen==1.47.0
rapidfuzz==3.8.1
rich==13.7.1
tqdm==4.66.4
tenacity==8.3.0
python-dotenv==1.0.1
typing-extensions==4.12.2
pytest==8.3.3
